Reorganizing components and cleaning up imports
