+++ tournament-game/src/components/tournament/BracketContext.js
@@ -1,15 +1,13 @@
 import React, {useState, createContext} from 'react';
-import {generateBracket, selectNodeRaw} from '../util';
+import {generateBracket, selectNodeRaw} from '../../util';
 
 export const BracketContext = createContext();
 
 export const BracketProvider = (props) => {
     
     const size = props.size;
-    // const size = parseInt(props.match.params.number);
     const [array, setArray] = useState(generateBracket(size));
     const selectNode = selectNodeRaw(array, setArray);
-    // const tree = renderTree(array, selectNode(array, setArray));
     const value = {array, selectNode}
