+++ tournament-game/src/components/tournament/Pairings.js
