+++ tournament-game/src/components/tournament/index.js
