+++ tournament-game/src/util/index.js
@@ -1,3 +1,2 @@
 export {default as generateBracket} from './generateBracket';
-export {default as renderTree} from './renderTree';
 export {default as selectNodeRaw} from './selectNodeRaw';
